lexer: structure and atoms
Support for structure elements of JSON, specifically '{}[],:' as well as
three atoms: 'true', 'false' and 'null'. These are tokenized in an
optimised way, because their starting characters are unique. I use this
pattern to detect the atoms and read the expected bytes (3 for true,
null and 4 for false) into a buffer, comparing the read bytes with the
expected characters inline (this is possible because r,u,e,a,l,s are all
valid ASCII chars and thus fit in a byte). This approach saves multiple
iterations of the tokenizing loop as well as multiple calls to
lexer.advance(), which in turn calls the lexer.r.ReadByte() method. Byte
comparison and skipping the manual switch entry for every byte we can
already expect should also save some runtime latency.